Major change! Redefine hash of a file in basename-mode.
Previously: H(basename ‖ U+001F ‖ content)
New: H(basename) ⭙ H(content)

Why? To be consistent with the definition of hashes of folders.
This actually allows to compute the empty mode equivalent report file
form the basename mode report file possible. But does it change the
relation of modes in the user interface? I thought about this. I don't
think so.

But it is funny that a folder with 3 files with distinct files has the
same hash if we rotate the names of files around. I don't think this is
a problem.