Remove all filepath interpretation from report file entries.
Why? The 'find' command interprets the file path to build the
corresponding tree. To represent the match, it needs to print
the original path. However, the “filepath interpretation” approach
means that “a//b” will be represented as “a/b” in the output.
This is annoying. We can simply get rid of this problem,
but considering it as a string split by ‘/’ or ‘\’.
This creates nodes with empty names, but this is okay.