Allow "" as filepath in report files.
This commit is related to commit a1ea20dc9152f7512849d45818d21f51d67ae9d0
Why do we want that? The issue is that all filesystem semantics have
been dropped in the previous commit. Therefore "." and "" is not
equivalent. Therefore "." and "" can be two separate entries.
I think it is a little bit awkward to read in report files,
but acceptable.