Add filepath separator to report file header.
This commit is related to commit a1ea20dc9152f7512849d45818d21f51d67ae9d0
Because of this previous commit, we lost the notion of a filepath
separator. Therefore we need to add it as metadata to the report file.
Any byte<128 can be used as separator to split a filepath into components.